requirements.txt
fastapi
firebase_admin
pydantic
python-dotenv
requests
starlette
email-validator